+++ guillotina/json/serialize_value.py
@@ -41,6 +41,11 @@ def string_converter(value):
     return str(value)
 
 
+@configure.value_serializer(bytes)
+def bytes_converter(value):
+    return str(value, encoding="utf-8")
+
+
 @configure.value_serializer(list)
 def list_converter(value):
     return list(map(json_compatible, value))